+++ src/components/CharacterList.vue
@@ -1,59 +1,106 @@
 <script setup lang="ts">
-import { onMounted } from 'vue';
+import { ComputedRef, computed, onMounted } from 'vue';
+
+import { useQuantityBasedOnWidth } from '@composables/useQuantityBasedOnWidth';
 import CardLoading from './CardLoading.vue';
 import CharacterCard from './CharacterCard.vue';
 
+import { Character } from '@/domain';
 import { useCharactersStore } from '@/infrastructure/store/characters';
 
 const store = useCharactersStore();
+const { totalOfCardsLoading, totalOfPagination } = useQuantityBasedOnWidth();
 
 onMounted(async () => {
     await store.getAllCharacters();
 });
 
+const listOfCharacters: ComputedRef<Character[]> = computed(() =>
+    !store.isSearching && !store.isLoading
+        ? store.charactersPerPage[store.pagination.currentPage]
+        : store.search.characters
+);
+
 const openCharacter = (id: number) => {
     console.warn('Abriu o personagem', id);
+};
+
+const changeThePage = async (page: number) => {
+    store.$patch({ pagination: { currentPage: page } });
+
+    if (!store.charactersPerPage[store.pagination.currentPage]) {
+        await store.getAllCharacters();
+    }
 };
 </script>
 
 <template>
     <section class="character-list">
-        <h4 class="character-list__qtd" data-list="qtd">{{ store.listOfCharacters.length }} Characters</h4>
+        <h4 v-if="!store.isSearching && !store.isLoading" class="character-list__qtd" data-list="results">
+            {{ store.pagination.results }} results
+        </h4>
         <main class="character-list__cards">
-            <CardLoading v-if="store.isLoading" />
             <CharacterCard
-                v-for="(character, index) in store.listOfCharacters"
-                v-else
+                v-for="(character, index) in listOfCharacters"
                 :id="Number(character.id)"
                 :key="index"
                 :name="character.name"
                 :species="character.species || ''"
                 :image="character.image"
-                alt-image="nada por aqui"
+                :alt-image="`${character.name} of specie ${character.species}`"
                 data-list="card"
                 @click="openCharacter(character.id)"
             />
+            <CardLoading v-if="store.isLoading" :quantity="totalOfCardsLoading" />
         </main>
 
-        <v-btn class="character-list__button" variant="tonal" data-list="button-load-more">Load more</v-btn>
+        <v-pagination
+            v-if="store.pagination.total > 0 && !store.isSearching"
+            v-model="store.pagination.currentPage"
+            :length="store.pagination.total"
+            :total-visible="totalOfPagination"
+            rounded="circle"
+            data-list="pagination"
+            variant="elevated"
+            active-color="meri-light"
+            @update:model-value="changeThePage"
+        ></v-pagination>
     </section>
 </template>
 
 <style lang="postcss" scoped>
 .character-list {
-    @apply px-10 flex flex-col justify-center items-center;
+    @apply px-10 
+        flex flex-col justify-center items-center;
 
     &__qtd {
-        @apply pb-6 font-semibold text-lg text-gray-500;
+        @apply pt-4 pb-6 
+            font-semibold text-lg text-gray-500 text-left
+            w-full;
     }
 
     &__cards {
-        @apply py-2 flex flex-row justify-start flex-wrap gap-2.5;
+        @apply py-2 
+            flex flex-row flex-wrap
+            w-full;
     }
+}
+</style>
 
-    &__button {
-        @apply rounded-2xl w-10/12 md:w-1/2 m-auto mt-6 bg-meri-mid text-white font-sans
-        text-lg;
-    }
+<style>
+.v-btn--disabled.v-btn--variant-elevated .v-btn__content {
+    font-weight: 700;
+    color: gray;
+}
+
+.v-pagination__prev .v-btn--disabled.v-btn--variant-elevated,
+.v-pagination__item .v-btn--disabled.v-btn--variant-elevated {
+    background-color: white;
+    color: white;
+}
+
+.v-pagination__prev .v-btn--disabled.v-btn--variant-elevated .v-btn__content,
+.v-pagination__item .v-btn--disabled.v-btn--variant-elevated .v-btn__content {
+    color: theme('colors.gray.300');
 }
 </style>