+++ src/views/PageHome.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
 import BaseHeader from '@components/BaseHeader.vue';
 import CharacterList from '@components/CharacterList.vue';
-import SearchActions from '@components/SearchActions.vue';
+import SearchAction from '@components/SearchActions.vue';
 </script>
 
 <template>
     <BaseHeader />
-    <SearchActions />
+    <SearchAction />
     <CharacterList />
 </template>