+++ tailwind.config.cjs
@@ -2,7 +2,14 @@
 module.exports = {
     content: ['./index.html', './src/**/*.{vue,ts,tsx}'],
     theme: {
-        extend: {},
+        extend: {
+            animation: {
+                'short-spin': 'spin 0.3s ease',
+            },
+            boxShadow: {
+                sprawled: '0px 10px 51px -5px rgba(183, 189, 193, 0.3)',
+            },
+        },
         colors: {
             transparent: 'transparent',
             white: '#FFFFFF',
@@ -32,6 +39,7 @@ module.exports = {
             },
             gray: {
                 100: '#f3f4f6',
+                150: '#eff3f6',
                 200: '#e5e7eb',
                 300: '#d1d5db',
                 400: '#a1a1aa',
@@ -43,13 +51,18 @@ module.exports = {
             },
         },
         screens: {
-            sm: '480px',
-            tablet: '640px',
-            md: '768px',
-            lg: '976px',
+            sm: { min: '480px', max: '767px' },
+            md: { min: '768px', max: '975px' },
+            lg: { min: '976px', max: '1439px' },
+            xl: '1440px',
+            mobile: '480px',
+            tablet: '768px',
             laptop: '1024px',
             desktop: '1280px',
-            xl: '1440px',
+        },
+        fontFamily: {
+            sans: ['Graphik', 'sans-serif'],
+            serif: ['Merriweather', 'serif'],
         },
     },
     plugins: [],